v0.12.4: docs, present-tense comments, and a 2x faster suite
Comments and docstrings across src/ and tests/ describe what the code
does, not what it once did. Rewritten rather than deleted: the reason a
rule exists is the valuable part, so "X used to happen, which was wrong"
becomes "X must not happen, because...". Reviewed by hand — grep cannot
tell "used to smuggle a denied binary" from "five modules used to spell
this out". CHANGELOG.md and AGENTS.md left alone by request.

docs/ gains eleven short how-to pages, linked from the README, and the
screenshot moves to docs/images/.

Tests: 207s -> 98s, ~5/s -> ~10/s, nothing removed.
- Session keys derive at scrypt's interactive cost suite-wide;
  test_crypto.py restores the shipped cost, since there the KDF is the
  subject. Cipher, blob format and round trip still exercised for real.
- The probe's give-up path is proven in 1.2s rather than 60.

Two real bugs found while profiling:
- StdioClient.close() blocked for as long as an unresponsive MCP server
  lived. It closed the pipes before ending the process, and a reader
  thread in `for line in process.stdout` holds that stream's lock, so
  the close waited on a read that would never return. 30s -> 2s.
- Seven /image argument-parsing tests were removed by accident in
  v0.12.1 along with a debug test; the suite stayed green because what
  remained passed. Restored.

+++ src/cobirb/flock/probe.py
@@ -93,7 +93,7 @@ def _timed_call(provider: Any, prompt: str) -> tuple[float, float]:
 
 
 def probe_concurrency(
-    provider: Any, *, timeout: int = DEFAULT_TIMEOUT, prompt: str = _PROMPT
+    provider: Any, *, timeout: float = DEFAULT_TIMEOUT, prompt: str = _PROMPT
 ) -> ProbeResult:
     """Find out whether ``provider``'s endpoint answers two requests at once.
